Activity Log: Add unavailability banner
In certain circumstances Rewind is unavailable and yet there is
something the customer can do about it. In these cases we want to
provide them a hint at what that might be and an opportunity to
act on that.

In this patch we're creting a banner to display in the Activity Log
to show when a Jetpack site is disconnected and also when VaultPress
is active but also able to transfer to Rewind.